ARCO-105: Move peer handler integration test
